src/backend/pyproject.toml
[project]
name = "aalec-quiz-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
